+++ app/serializers/related_article_serializer.rb
@@ -1,0 +1,11 @@
+# Serializes related articles, which represent links between
+# this manuscript and others, published or not.
+class RelatedArticleSerializer < ActiveModel::Serializer
+  attributes :id,
+             :paper_id,
+             :linked_doi,
+             :linked_title,
+             :additional_info,
+             :send_manuscripts_together,
+             :send_link_to_apex
+end