+++ engines/tahi_standard_tasks/client/app/models/related-articles-task.js
@@ -1,0 +1,5 @@
+import Task from 'tahi/models/task';
+
+export default Task.extend({
+  qualifiedType: 'TahiStandardTasks::RelatedArticlesTask'
+});